@@ -15,19 +15,19 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 0.1.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.1.1 <=4.1.1"
 
 dependencies:
-  platform: ^3.1.0
+  platform: ^3.1.2
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
@@ -38,24 +38,24 @@ dev_dependencies:
 
   # For information on the generic Dart part of this file, see the
   # following page: https://dart.dev/tools/pub/pubspec
-  font_awesome_flutter: ^8.11.0
-  sqflite: ^1.3.0
-  fl_chart: ^0.30.0
-  flutter_local_notifications: ^5.0.0
-  flutter_native_timezone: ^1.0.10
-  flutter_bloc: ^3.2.0
-  easy_localization: ^3.0.0
-  device_info: ^2.0.0
-  table_calendar: ^2.3.3
-  image_picker: ^0.7.0
-  crypto: ^3.0.1
+  font_awesome_flutter: ^10.5.0
+  sqflite: ^2.3.0
+  fl_chart: ^0.63.0
+  flutter_local_notifications: ^15.1.1
+  flutter_native_timezone: ^2.0.0
+  flutter_bloc: ^8.1.3
+  easy_localization: ^3.0.3
+  device_info: ^2.0.3
+  table_calendar: ^3.0.9
+  image_picker: ^1.0.4
+  crypto: ^3.0.3
   hex: ^0.2.0
-  dio: ^4.0.0
-  modal_progress_hud_nsn: ^0.1.0-nullsafety-1
-  syncfusion_flutter_pdf: ^18.4.49-beta
-  open_file: ^3.2.1
+  dio: ^5.3.3
+  modal_progress_hud_nsn: ^0.4.0
+  syncfusion_flutter_pdf: ^23.1.38
+  open_file: ^3.3.2
   string_similarity: ^2.0.0
-  dcdg: ^3.1.1
+  dcdg: ^4.1.0
 
 
 
