@@ -2,15 +2,18 @@
 import 'package:flutter/material.dart';
 
 class ReusableCard extends StatelessWidget {
-  ReusableCard(
-      {@required this.color,
-      @required this.cardChild,
-      this.onPressed,
-      this.onLongPressed});
+  ReusableCard({
+    required this.color,
+    required this.cardChild,
+    this.onPressed,
+    this.onLongPressed,
+  });
+
   final Color color;
   final Widget cardChild;
-  final Function onPressed;
-  final Function onLongPressed;
+  final VoidCallback? onPressed;
+  final VoidCallback? onLongPressed;
+
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
