@@ -16,15 +16,13 @@ class AlarmDataBaseProvider {
   AlarmDataBaseProvider._();
   static final AlarmDataBaseProvider db = AlarmDataBaseProvider._();
 
-  Database _database;
+  Database? _database;
 
-  // 返回数据库，一般情况无需更改
+  // Returns the database, generally no need to change
   Future<Database> get database async {
-    if (_database != null) {
-      return _database;
-    }
+    if (_database != null) return _database!;
     _database = await createDatabase();
-    return _database;
+    return _database!;
   }
 
   // 创建数据库参数，一般情况无需更改
