@@ -13,15 +13,13 @@ class SleepDataBaseProvider {
   SleepDataBaseProvider._();
   static final SleepDataBaseProvider db = SleepDataBaseProvider._();
 
-  Database _database;
+  Database? _database;
 
-  //get database在flutter中为getter的写法
+  // Returns the database, generally no need to change
   Future<Database> get database async {
-    if (_database != null) {
-      return _database;
-    }
+    if (_database != null) return _database!;
     _database = await createDatabase();
-    return _database;
+    return _database!;
   }
 
   Future<Database> createDatabase() async {
