@@ -22,7 +22,7 @@ class _BloodPressureState extends State<BloodPressure> {
   int dbPressure = 75; //低压
   int heartRate = 60;
 
-  TextEditingController _voiceInputController;
+  late TextEditingController _voiceInputController;
   String _voiceInput = '';
 
   _pressureWarningDialog(BuildContext context) {
@@ -77,7 +77,7 @@ class _BloodPressureState extends State<BloodPressure> {
     RegExp regExp = new RegExp(r'([0-9]{3}|[0-9]{2})');
     Iterable<Match> matches = regExp.allMatches(text);
     for (Match m in matches) {
-      int match = int.parse(m[0]);
+      int match = int.parse(m[0]!);
       bpData.add(match);
     }
     if (bpData.isEmpty) {
@@ -465,7 +465,7 @@ class _BloodPressureState extends State<BloodPressure> {
                   ),
                 ),
                 ButtonButton(
-                  onTap: () {
+                  onTap: () async {
                     if (_voiceInput != "") {
                       print(_voiceInput);
                       List bpList = checkBP(_voiceInput);
@@ -534,7 +534,7 @@ class _BloodPressureState extends State<BloodPressure> {
                         }
                       } else {
                         // 匹配失败
-                        return showDialog<void>(
+                        return await showDialog(
                             context: context,
                             barrierDismissible: false, // user must tap button!
                             builder: (BuildContext context) {
